+++ lib/origin/client/origin_server_connection.py
@@ -1,13 +1,13 @@
 import json
 import sys
-from origin import data_types, config, timestamp
+from origin import data_types, timestamp
 import struct
 import ctypes
 
 # returns string and size tuple
-def makeFormatString(keyOrder, records):
+def makeFormatString(config, keyOrder, records):
     try:
-        tsType = config["timestamp_type"]
+        tsType = config.get('Server',"timestamp_type")
     except KeyError:
         tsType = "uint"
     tsSize = data_types[tsType]["size"]
@@ -21,7 +21,8 @@ def makeFormatString(keyOrder, records):
 
 
 class server_connection:
-    def __init__(self,stream,streamID,keyOrder,format,records,context,socket):
+    def __init__(self, config, stream, streamID, keyOrder, format, records, context, socket):
+        self.config = config
         self.stream = stream
         self.streamID = streamID
         self.keyOrder = keyOrder
@@ -35,24 +36,25 @@ class server_connection:
         if keyOrder is None:
             self.format_string, self.data_size = (None, None)
         else:
-            self.format_string, self.data_size = makeFormatString(keyOrder,records)
+            self.format_string, self.data_size = makeFormatString(self.config,keyOrder,records)
 
     def send(self,**kwargs):
         msgData = [ self.streamID ]
         try:
             msgData.append(kwargs[timestamp])
         except KeyError:
-            # 0 value timestamp means timestamp at server
+            #print "No timestamp specified, server will timestamp on arrival"
             msgData.append(0)
         if self.format is None:
             for k in self.keyOrder:
                 msgData.append(kwargs[k])
             self.socket.send( self.formatRecord(msgData) )
         elif self.format == "json":
+	    msgData[0] = self.stream
             msgMap = {}
             for k in kwargs.keys():
-                if k != config.timestamp:
-                    msgMap[k] = kwqrgs[k]
+                if k != timestamp:
+                    msgMap[k] = kwargs[k]
             msgData.append(msgMap)
             self.socket.send(json.dumps(msgData))