+++ lib/origin/server/origin_filesystem_destination.py
@@ -1,87 +1,145 @@
-import os
 import json
+from origin.server import destination
+from origin import data_types, timestamp
+import os
+import numpy as np
+import sys, traceback
 
-from origin.server import measurement_validation
-from origin.server import template_validation
+def getDirectoryList(dir):
+    return [ d for d in os.listdir(dir) if os.path.isdir(os.path.join(dir,d)) ]
 
-class filesystem_destination:
-    def __init__(self,cachedir):
-        if not os.path.exists(cachedir):
-            os.mkdir(cachedir)
-        self.cachedir = cachedir
+def getCurrentStreamVersion(config,stream):
+    stream_path = config.get('FileSystem','data_path')
+    stream_path = os.path.join( config.get('Server', "var_path"), stream_path, stream )
+    with open(os.path.join(stream_path,'currentVersion.txt'), 'r') as f:
+        version_dir = f.read().strip()
+    return os.path.join(stream_path,version_dir)
 
-        streamdef = "%s/stream-def"%(cachedir)
-        if not os.path.exists(streamdef):
-            os.mkdir(streamdef)
-        self.streamdef = streamdef
+class filesystem_destination(destination):
+    def connect(self):
+        self.data_path = self.config.get('FileSystem','data_path')
+        self.data_path = os.path.join( self.config.get('Server', "var_path"), self.data_path )
+        self.info_file = os.path.join( self.data_path, self.config.get('FileSystem','info_file') )
+        if not os.path.exists(self.data_path):
+            os.makedirs(self.data_path)
+            self.logger.info("Creating data directory at: " + self.data_path)
 
-        streamdata = "%s/stream-data"%(cachedir)
-        if not os.path.exists(streamdata):
-            os.mkdir(streamdata)
-        self.streamdata = streamdata
-
-        knownStreamDirs = os.listdir(streamdef)
+    def readStreamDefTable(self):
         knownStreams = {}
-        for k in knownStreamDirs:
-            infile = open("%s/stream-def/%s"%(cachedir,k))
-            streamTemplate = {}
-            for line in infile:
-                tokens = line.rstrip().lstrip().split()
-                streamTemplate[tokens[0]] = tokens[1] 
+        knownStreamVersions = {}
+        try:
+            with open(self.info_file, 'r') as f:
+                json_data = f.read()
+            knownStreams = json.loads(json_data)
+        except IOError:
+            self.logger.debug("Info file `{}` not found".format(self.info_file))
+            knownStreams = {}
 
-            print "Found known stream '",k,"'"
-            knownStreams[k] = streamTemplate
-        self.knownStreams = knownStreams
-        self.knownStreamDirs = knownStreamDirs
-        
-    def registerStream(self,stream,template):
-        result = None
-        resultText = None
-        
-        if stream not in self.knownStreams.keys():
-            print "Uknown stream",stream
-            print "Will register it."
-            outfile = open("%s/%s"%(self.streamdef,stream),"w")
+        for stream in knownStreams:
+            knownStreamVersions[stream] = knownStreams[stream]['definition']
 
-            outdir = "%s/%s"%(self.streamdata,stream)
-            if not os.path.exists(outdir):
-                os.mkdir(outdir)
-            
-            self.knownStreams[stream] = template
+        self.knownStreams=knownStreams
+        self.knownStreamVersions=knownStreamVersions
+        self.print_stream_info()
 
-            for name in template.keys():
-                print >> outfile,name,template[name]
+    def createNewStreamDestination(self,stream_obj):
+        stream = stream_obj["stream"]
+        version = stream_obj["version"]
+        stream_path = os.path.join(os.path.join(self.data_path,stream))
+        if version == 1:    # create a new stream group under root
+            os.mkdir(stream_path)
 
-        if template_validation(template,self.knownStreams[stream]):
-            print "it appears legitimate"
-            result = 0
-            resultText = ""
-        else:
-            print "it doesn't validate"
-            result = 1
-            resultText = "Problem with template compared to old template"
-                
-        return (result,resultText)
+        # create a new subgroup for this instance of the current stream
+        version_dir = stream + '_' + str(version)
+        stream_ver = os.path.join( stream_path, version_dir )
+        os.mkdir( os.path.join(stream_path, version_dir) )
+        # point currentVersion.txt file at most current version
+        with open( os.path.join(stream_path, 'currentVersion.txt'), 'w' ) as f:
+            f.write(version_dir)
+        # update the main info file
+        with open(self.info_file, 'w') as json_data:
+            json_data.write(json.dumps(self.knownStreams))
+        # create the stream field definition dict file
+        with open( os.path.join(stream_path, version_dir, 'definition.json'), 'w' ) as f:
+            f.write( json.dumps(stream_obj['definition']) )
+        return stream_obj['id']
 
-    def measurement(self,measurementTime,stream,measurements):
-        if stream not in self.knownStreams.keys():
-            print "trying to add a measurement to data on an unknown stream"
-            return (1,"Unknown stream")
+    def insertMeasurement(self,stream,measurements):
+        current_stream_version = getCurrentStreamVersion(self.config,stream)
+        for field in measurements:
+            with open( os.path.join(current_stream_version, field), 'a' ) as f:
+                f.write( str(measurements[field]) + '\n')
 
-        if not measurement_validation(measurements,self.knownStreams[stream]):
-            print "Measurement didn't validate against the pre-determined format"
-            return (1,"Invalid measurements against schema")
-        outdir = self.streamdata
+    # read stream data from storage between the timestamps given by time = [start,stop]
+    # only takes timestamps in seconds
+    def getRawStreamData(self,stream,start=None,stop=None,definition=None):
+        start, stop = self.validateTimeRange(start,stop)
+        # get data from buffer
+        try:
+            current_stream_version = getCurrentStreamVersion(self.config,stream)
+        except IOError:
+            self.logger.debug("Read requested on empty dataset. stream: {}", stream)
+            raise IndexError
+        if definition is None:
+            definition = self.knownStreamVersions[stream]
 
-        outfile = open("%s/%s/%d"%(outdir,stream,measurementTime),'a')
-        for m in measurements.keys():
-            print >> outfile,m,measurements[m]
-        outfile.close()
+        start_idx = None
+        stop_idx = None
+        try:
+            tsfile = os.path.join(current_stream_version,timestamp)
+            f = open(tsfile,'r')
+            i = 0
+            data = { timestamp : [] }
+            dtype = data_types[self.config.get('Server','timestamp_type')]['type']
+            while True:
+                x = f.readline()
+                x = x.strip()
+                if not x: break
+                x = dtype(x) 
+                if (start_idx is None) and (x >= start):
+                    start_idx = i
+                if (start_idx is not None) and (x >= stop):
+                    stop_idx = i
+                    break
+                if start_idx is not None:
+                    data[timestamp].append(x)
+                i += 1
+            if (start_idx is not None) and (stop_idx is None):
+                stop_idx = i
 
-        result = 0
-        resultText = ""
-        return (result,resultText)
+        except ValueError:
+            self.logger.error("Error casting timestamp to type long. x=`{}` in {}".format(x, tsfile))
+        except:
+            self.logger.error("Unexpected exception reading from data file. Message code:")
+            self.logger.error(traceback.print_exc(file=sys.stdout))
+        finally:
+            f.close()
 
+        if (start_idx is None) or (stop_idx is None):
+            msg = "error in indexing (index start, index stop): ({},{})"
+            self.logger.warning(msg.format(start_idx,stop_idx))
+            raise IndexError
 
-            
-        
+        self.logger.debug((start_idx,stop_idx))
+        for field in definition:
+            data[field] = []
+            dtype = self.knownStreamVersions[stream][field]['type']
+            type_cast = data_types[dtype]['type']
+            try:
+                f = open(os.path.join(current_stream_version,field),'r')
+                i = 0
+                while i <= stop_idx:
+                    x = f.readline()
+                    if i >= start_idx:
+                        x = type_cast(x.strip())
+                        data[field].append(x)
+                    i += 1
+            except ValueError:
+                pass # reached EOF
+            except:
+                self.logger.error("Unexpected exception reading from data file. Message code:")
+                self.logger.error(traceback.print_exc(file=sys.stdout))
+                raise IndexError
+            finally:
+                f.close()
+        return data