+++ lib/origin/server/origin_hdf5_destination.py
@@ -1,55 +1,54 @@
 import h5py
 import json
 from origin.server import destination
-from origin import data_types, config, timestamp
+from origin import data_types, timestamp
 import os
 import numpy as np
 
 class hdf5_destination(destination):
     def connect(self):
-        data_dir = config['data_path']
+        data_dir = self.config.get('HDF5','data_path')
+        if not os.path.exists( data_dir ):
+            data_dir = os.path.join(self.config.get("Server",'var_path'), data_dir)
+
         if not os.path.exists(data_dir):
             os.makedirs(data_dir)
-            self.logger.info("Creating data directory at: " + config['data_path'])
+            self.logger.info("Creating data directory at: " + self.config.get('HDF5','data_path'))
+
+        f = os.path.join( data_dir, self.config.get('HDF5','data_file') )
         try:
-            self.hdf5_file = h5py.File(config['data_file'], 'r+')
-            self.logger.info("Opened data file: {}".format(config['data_file']))
+            self.hdf5_file = h5py.File(f, 'r+')
+            self.logger.info("Opened data file: {}".format(f))
         except IOError:
             try:
-                self.hdf5_file = h5py.File(config['data_file'], 'w')
-                self.logger.info("New data file: {}".format(config['data_file']))
+                self.hdf5_file = h5py.File(f, 'w')
+                self.logger.info("New data file: {}".format(f))
             except IOError:
-                self.logger.error("Unable to create data file: " + config['data_file'])
-
+                self.logger.error("Unable to create data file: {}".format(f))
 
     def readStreamDefTable(self):
+        knownStreams = {}
+        knownStreamVersions = {}
         try:
-            self.knownStreamVersions = json.loads(self.hdf5_file.attrs['knownStreamVersions'])
+            knownStreams = json.loads(self.hdf5_file.attrs['knownStreams'])
         except KeyError:
-            self.logger.debug("knownStreamVersions attribute not found") 
-            self.knownStreamVersions = {}
+            self.logger.debug("knownStreams attribute not found") 
+            knownStreams = {}
 
-        self.knownStreams = {}
-        for stream in self.knownStreamVersions:
-            if stream in self.hdf5_file:
-                current_stream_version = self.hdf5_file[stream].attrs['currentVersion']
-                definition = self.hdf5_file[current_stream_version].attrs['definition']
-                self.knownStreams[stream] = json.loads(definition)
-            else:
-                self.logger.error("Stream '{}' found in stream list, but is not a group")
+        for stream in knownStreams:
+            knownStreamVersions[stream] = knownStreams[stream]['definition']
 
-        for stream in self.knownStreams:
-            print "="*10, " {} ".format(stream), "="*10
-            for field_name in self.knownStreamVersions[stream]:
-                print "  Field: %s (%s)"%(field_name,self.knownStreamVersions[stream][field_name])
+        self.knownStreams=knownStreams
+        self.knownStreamVersions=knownStreamVersions
+        self.print_stream_info()
 
-    def createNewStream(self,stream,version,template,keyOrder):
+    def createNewStreamDestination(self,stream_obj):
+        stream = stream_obj['stream']
+        version = stream_obj['version']
         if version == 1:    # create a new stream group under root
             stream_group = self.hdf5_file.create_group(stream)
-            streamID = len(self.hdf5_file)
         else:
             stream_group = self.hdf5_file[stream]
-            streamID = self.knownStreamVersions[stream]["id"]
         # create a new subgroup for this instance of the current stream
         stream_ver = stream_group.create_group( stream + '_' + str(version) )
 
@@ -59,55 +58,48 @@ class hdf5_destination(destination):
 
         # data sets for each field plus the timestamp
         # also make a buffer dataset for each field a as a pseudo circular buffer
-        chunksize=(config['hdf5_chunksize'],) 
-        buff_size = (config['hdf5_chunksize'],)
-        print buff_size
+        chunksize=(self.config.getint('HDF5','chunksize'),) 
+        buff_size = chunksize
+        #print buff_size
+        tstype = self.config.get('Server','timestamp_type')
+        compression = self.config.get('HDF5','compression')
         stream_ver.create_dataset(
                 timestamp
                 , chunksize
                 , maxshape = (None,)
-                , dtype=data_types[config['timestamp_type']]['numpy']
+                , dtype=data_types[tstype]['numpy']
                 , chunks=chunksize
-                , compression=config['hdf5_compression']
+                , compression=compression
         )
         stream_ver.create_dataset(
                 timestamp + "_buffer"
                 , buff_size
                 , maxshape = buff_size
-                , dtype=data_types[config['timestamp_type']]['numpy']
+                , dtype=data_types[tstype]['numpy']
                 , chunks=chunksize
         )
+        template = stream_obj['definition']
         for field in template:
+            dt = data_types[template[field]['type']]['numpy']
             stream_ver.create_dataset(
                     field
                     , chunksize
                     , maxshape = (None,)
-                    , dtype=data_types[template[field]]['numpy']
+                    , dtype=dt
                     , chunks=chunksize
-                    , compression=config['hdf5_compression']
+                    , compression=compression
             )
             stream_ver.create_dataset(
                     field + "_buffer"
                     , buff_size
                     , maxshape = buff_size
-                    , dtype=data_types[template[field]]['numpy']
+                    , dtype=dt
                     , chunks=chunksize
             )
     
-        # update the stream inventory, in memory and on disk
-        self.knownStreamVersions[stream] = {
-                "version"   : version,
-                "id"        : streamID,
-                "keyOrder"  : keyOrder,
-                "formatStr" : self.formatString(template,keyOrder),
-        }
-        self.hdf5_file.attrs['knownStreamVersions'] = json.dumps(self.knownStreamVersions)
-        # create the stream field definition dict
-        definition = {}
-        for i, key in enumerate(keyOrder):
-            definition[key] = { "type": template[key], "keyIndex": i }
-        stream_ver.attrs['definition'] = json.dumps(definition)
-        return streamID
+        self.hdf5_file.attrs['knownStreams'] = json.dumps(self.knownStreams)
+        stream_ver.attrs['definition'] = json.dumps(self.knownStreamVersions[stream])
+        return stream_obj['id']
 
     def insertMeasurement(self,stream,measurements):
         dgroup = self.hdf5_file[self.hdf5_file[stream].attrs['currentVersion']]
@@ -118,10 +110,10 @@ class hdf5_destination(destination):
             if (row_count_buffer == buffer_size):
                 self.logger.debug("Buffer is full. Moving completed chunk to archive and wrapping pointer around.")
                 length = dgroup[timestamp].shape[0]
-                chunksize=config['hdf5_chunksize'] 
+                chunksize=self.config.getint('HDF5','chunksize')
                 for field in measurements:
                     dgroup[field][row_count:] = dgroup[field+'_buffer']
-                    dgroup[field].resize((length+config['hdf5_chunksize'],))
+                    dgroup[field].resize((length+chunksize,))
                 row_count += buffer_size
                 row_count_buffer = 0
         else:
@@ -130,11 +122,10 @@ class hdf5_destination(destination):
 
         dgroup.attrs['row_count'] = row_count # move pointer for next entry
         dgroup.attrs['row_count_buffer'] = row_count_buffer # move pointer for next entry
-        #self.logger.debug("Stream `{}` current row pointer: {}".format(stream,row_count))
-        #self.logger.debug("Stream `{}` current row buffer pointer: {}".format(stream,row_count_buffer))
 
-        #self.logger.debug("Datasets `{}.*` shape: {}".format(stream,dgroup[timestamp].shape))
         for field in measurements:
+            if (field != timestamp) and (self.knownStreamVersions[stream][field]['type'] == "string"):
+                measurements[field] = measurements[field].encode("ascii","ignore")
             dgroup[field+"_buffer"][row_count_buffer] = measurements[field]
 
     # read stream data from storage between the timestamps given by time = [start,stop]
@@ -145,7 +136,7 @@ class hdf5_destination(destination):
         # get data from buffer
         stream_group = self.hdf5_file[self.hdf5_file[stream].attrs['currentVersion']]
         if definition is None:
-            definition = self.knownStreams[stream]
+            definition = self.knownStreamVersions[stream]
 
         # read ring buffers and pointers in case the pointer advances during read
         raw_data = {}
@@ -213,10 +204,6 @@ class hdf5_destination(destination):
             data[field] = raw_data[field][idx_start:idx_stop].tolist() # json method cant handle numpy array
         return data
         
-    # read stream.field data from storage between the timestamps given by time = [start,stop]
-    def getRawStreamFieldData(self,stream,field,start=None,stop=None):
-        return self.getRawStreamData(stream=stream,start=start,stop=stop, definition={field:''}) # send dummy dict with single field
-
     # get raw_data in range from the archived data
     def getArchivedStreamData(self,stream_group,start,stop,buffer_data,definition):
         time_dset = stream_group[timestamp]
@@ -226,7 +213,7 @@ class hdf5_destination(destination):
         old_data[timestamp] = np.zeros(row_pointer)
         for field in definition:
             old_data[field] = np.zeros(row_pointer) 
-        chunksize = config['hdf5_chunksize']
+        chunksize = self.config.getint('HDF5','chunksize')
         row_pointer -= chunksize
         while row_pointer >= 0:
             chunk = time_dset[row_pointer:row_pointer+chunksize]