+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Node registration with ordered, raw string format
 - Binary datapacket support, with token identification
 - Backend storage with HDF5
+- Backend storage with raw filesystem files
 - Supported datatypes: [u]int{8,16,32,64}, float (32b), double (64b), string (10 elements in hdf5)
 - Server adds timestamps if no timestamp is sent or timestamp is 0
 - Added data access API function to server
@@ -14,6 +15,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Timestamp default datatype changed to uint64 from int32
 - Default storage changed to HDF5
+- Dictionary config file to python config file
 
 ## 0.1.0 - 2016-06-07
 ### Added